Mainly fix the mistake for random solver
Previously treat the every 'zz' as random integration,that is sqrt time step, but in fact only rho(that is zz[n/2]) is
the gaussian random number, followings are all deterministic and should be integrated with full 'dt' time step. This has
been added to both Euler and vv algorithem, the former has been commented out.