Add heat current for diatomic
So mainly diatomic*withbaths.py and diatomic*Gaussian.py is added. The former is the one I am working on as the test-bed
for heat current of Debye spectrum, the latter is just soft of "double check" file to confirm White noise
random-calculated current equals to "force" calculated power.

Though might not have to mention, if look at the comment-out "if" condition lines in the *Gaussian file, one may
notice, the time step consistency is extremely important to integrate random-bath-noise heat current. With just one step
off, desired results can not be simply recovered.